refactor(context-manager): move methods/ to context-manager level
methods/ contains reduction primitives (truncate, summarize) that are
not specific to the offload strategy — they belong as siblings of
strategies/, not nested inside it.
